adding example

+++ src/codexion.h
@@ -39,7 +39,7 @@ typedef struct s_coder
 	pthread_t	coder;
 	int			coder_id;
 	int			number_of_compiles;
-	int			last_compile_start;
+	long		last_compile_start;
 	t_input		*data;
 }				t_coder;
